is is ok??

+++ answermethod.c
@@ -28,15 +28,13 @@ void send_ok(DBusMessage * msg, DBusConnection * conn)
 
 	// add the arguments to the reply
 	dbus_message_iter_init_append(reply, &args);
-	if (!dbus_message_iter_append_basic(&args, DBUS_TYPE_INT32, &ok)) 
-	{
-		fprintf(stderr, "Out Of Memory!\n");
+	if (!dbus_message_iter_append_basic(&args, DBUS_TYPE_INT32, &ok)) {
+		panic("%s", "Out Of Memory!\n");
 		exit(1);
 	}
 	// send the reply && flush the connection
-	if (!dbus_connection_send(conn, reply, &serial)) 
-	{
-		fprintf(stderr, "Out Of Memory!\n");
+	if (!dbus_connection_send(conn, reply, &serial)) {
+		panic("%s", "Out Of Memory!\n");
 		exit(1);
 	}
 	dbus_connection_flush(conn);
@@ -53,36 +51,32 @@ void start(DBusMessage * msg, DBusConnection * conn)
 
 	DBusMessageIter args;
 	char *parmList[] = { "/usr/bin/ffmpeg", "-f", "video4linux2", "-s" };
-	if (status == 0) 
-	{
-		panic("%s", "already started \n");
+	if (status == 0) {
+		debug(1, "%s", "already started \n");
 		send_ok(msg, conn);
 		return;
 	}
+	args = args;
 // create the pipe
-	if ((pipe(pfderr) == -1) || (pipe(pfdout) == -1)) 
-	{
-		printf("pipe failed\n");
+	if ((pipe(pfderr) == -1) || (pipe(pfdout) == -1)) {
+		panic("%s", "pipe failed\n");
 		return;
 	}
 	status = 0;
-	printf("video started \n");
-	args = args;
+	debug(1, "%s", "video started \n");
 
 	/* create the child */
-	if ((pid = fork()) < 0) 
-	{
-		printf("fork failed\n");
+	if ((pid = fork()) < 0) {
+		panic("%s", "fork failed\n");
 		return;
 	}
 
-	if (pid == 0) 
-	{
+	if (pid == 0) {
 		/* child */
 		close(pfderr[0]);	/* close the unused read side */
 		close(pfdout[0]);
 		dup2(pfderr[1], 2);	/* connect the write side with stderr */
-		//      dup2(pfdout[1], 1);     /* connect the write side with stdout */
+		dup2(pfdout[1], 1);	/* connect the write side with stdout */
 		close(pfderr[1]);	/* close the write side */
 		close(pfdout[1]);
 		sprintf(carac, "%d", width);
@@ -110,11 +104,11 @@ void start(DBusMessage * msg, DBusConnection * conn)
 		parmList[15] = strdup(temp);
 		parmList[16] = NULL;
 		execvp("ffmpeg", parmList);
-		printf("Return not expected. Must be an execvp() error.\\n");
+		panic("%s",
+		      "Return not expected. Must be an execvp() error.\\n");
 		return;
 
-	} else 
-	{
+	} else {
 		/* parent */
 		send_ok(msg, conn);
 		return;
@@ -125,9 +119,10 @@ void start(DBusMessage * msg, DBusConnection * conn)
 
 void stop(DBusMessage * msg, DBusConnection * conn)
 {
-	kill(pid, SIGKILL);
-	printf("it's killed \n");
-	printf("\n");
+	if (status == 0) {
+		kill(pid, SIGKILL);
+		debug(1, "%s", "it's killed \n");
+	}
 	// create a reply from the message
 	send_ok(msg, conn);
 	status = 1;
@@ -138,20 +133,20 @@ void image_size_set(DBusMessage * msg, DBusConnection * conn)
 {
 	DBusMessageIter args;
 	// read the arguments
-	if (!dbus_message_iter_init(msg, &args))
-		fprintf(stderr, "Message has no arguments!\n");
-	else if (DBUS_TYPE_INT32 != dbus_message_iter_get_arg_type(&args))
-		fprintf(stderr, "Argument is not integer!\n");
+	if (!dbus_message_iter_init(msg, &args)) {
+		panic("%s", "Message has no arguments!\n");
+	} else if (DBUS_TYPE_INT32 != dbus_message_iter_get_arg_type(&args))
+		panic("%s", "Argument is not integer!\n");
 	dbus_message_iter_get_basic(&args, &width);
-	if (!dbus_message_iter_next(&args))
-		fprintf(stderr, "Message has too few arguments!\n");
-	else if (DBUS_TYPE_INT32 != dbus_message_iter_get_arg_type(&args))
-		fprintf(stderr, "Argument is not integer!\n");
-	else
+	if (!dbus_message_iter_next(&args)) {
+		panic("%s", "Message has too few arguments!\n");
+	} else if (DBUS_TYPE_INT32 != dbus_message_iter_get_arg_type(&args)) {
+		panic("%s", "Argument is not integer!\n");
+	} else
 		dbus_message_iter_get_basic(&args, &height);
 
-	printf("image width set:  %d\n", width);
-	printf("image height set:  %d\n", height);
+	debug(1, "%d", width);
+	debug(1, "%d", height);
 	send_ok(msg, conn);
 }
 
@@ -166,20 +161,17 @@ void image_size_get(DBusMessage * msg, DBusConnection * conn)
 
 	// add the arguments to the reply
 	dbus_message_iter_init_append(reply, &args);
-	if (!dbus_message_iter_append_basic(&args, DBUS_TYPE_INT32, &width)) 
-	{
-		fprintf(stderr, "Out Of Memory!\n");
+	if (!dbus_message_iter_append_basic(&args, DBUS_TYPE_INT32, &width)) {
+		panic("%s", "Out Of Memory!\n");
 		exit(1);
 	}
-	if (!dbus_message_iter_append_basic(&args, DBUS_TYPE_INT32, &height)) 
-	{
-		fprintf(stderr, "Out Of Memory!\n");
+	if (!dbus_message_iter_append_basic(&args, DBUS_TYPE_INT32, &height)) {
+		panic("%s", "Out Of Memory!\n");
 		exit(1);
 	}
 	// send the reply && flush the connection
-	if (!dbus_connection_send(conn, reply, &serial)) 
-	{
-		fprintf(stderr, "Out Of Memory!\n");
+	if (!dbus_connection_send(conn, reply, &serial)) {
+		panic("%s", "Out Of Memory!\n");
 		exit(1);
 	}
 	dbus_connection_flush(conn);
@@ -192,12 +184,12 @@ void image_rate_set(DBusMessage * msg, DBusConnection * conn)
 {
 	DBusMessageIter args;
 	// read the arguments
-	if (!dbus_message_iter_init(msg, &args))
-		fprintf(stderr, "Message has no arguments!\n");
-	else if (DBUS_TYPE_INT32 != dbus_message_iter_get_arg_type(&args))
-		fprintf(stderr, "Argument is not integer!\n");
+	if (!dbus_message_iter_init(msg, &args)) {
+		panic("%s", "Message has no arguments!\n");
+	} else if (DBUS_TYPE_INT32 != dbus_message_iter_get_arg_type(&args))
+		panic("%s", "Argument is not integer!\n");
 	dbus_message_iter_get_basic(&args, &imrate);
-	printf("image rate set:  %d\n", imrate);
+	debug(1, "%d", imrate);
 	send_ok(msg, conn);
 
 }
@@ -213,15 +205,13 @@ void image_rate_get(DBusMessage * msg, DBusConnection * conn)
 
 	// add the arguments to the reply
 	dbus_message_iter_init_append(reply, &args);
-	if (!dbus_message_iter_append_basic(&args, DBUS_TYPE_INT32, &imrate)) 
-	{
-		fprintf(stderr, "Out Of Memory!\n");
+	if (!dbus_message_iter_append_basic(&args, DBUS_TYPE_INT32, &imrate)) {
+		panic("%s", "Out Of Memory!\n");
 		exit(1);
 	}
 	// send the reply && flush the connection
-	if (!dbus_connection_send(conn, reply, &serial)) 
-	{
-		fprintf(stderr, "Out Of Memory!\n");
+	if (!dbus_connection_send(conn, reply, &serial)) {
+		panic("%s", "Out Of Memory!\n");
 		exit(1);
 	}
 	dbus_connection_flush(conn);
@@ -235,12 +225,13 @@ void bit_rate_set(DBusMessage * msg, DBusConnection * conn)
 {
 	DBusMessageIter args;
 	// read the arguments
-	if (!dbus_message_iter_init(msg, &args))
-		fprintf(stderr, "Message has no arguments!\n");
-	else if (DBUS_TYPE_INT32 != dbus_message_iter_get_arg_type(&args))
-		fprintf(stderr, "Argument is not integer!\n");
+	if (!dbus_message_iter_init(msg, &args)) {
+		panic("%s", "Message has no arguments!\n");
+	} else if (DBUS_TYPE_INT32 != dbus_message_iter_get_arg_type(&args)) {
+		panic("%s", "Argument is not integer!\n");
+	}
 	dbus_message_iter_get_basic(&args, &bitrate);
-	printf("Bit rate set:  %d\n", bitrate);
+	debug(1, "%d", bitrate);
 	send_ok(msg, conn);
 }
 
@@ -255,15 +246,13 @@ void bit_rate_get(DBusMessage * msg, DBusConnection * conn)
 
 	// add the arguments to the reply
 	dbus_message_iter_init_append(reply, &args);
-	if (!dbus_message_iter_append_basic(&args, DBUS_TYPE_INT32, &bitrate)) 
-	{
-		fprintf(stderr, "Out Of Memory!\n");
+	if (!dbus_message_iter_append_basic(&args, DBUS_TYPE_INT32, &bitrate)) {
+		panic("%s", "Out Of Memory!\n");
 		exit(1);
 	}
 	// send the reply && flush the connection
-	if (!dbus_connection_send(conn, reply, &serial)) 
-	{
-		fprintf(stderr, "Out Of Memory!\n");
+	if (!dbus_connection_send(conn, reply, &serial)) {
+		panic("%s", "Out Of Memory!\n");
 		exit(1);
 	}
 	dbus_connection_flush(conn);
@@ -277,12 +266,12 @@ void qscale_set(DBusMessage * msg, DBusConnection * conn)
 {
 	DBusMessageIter args;
 	// read the arguments
-	if (!dbus_message_iter_init(msg, &args))
-		fprintf(stderr, "Message has no arguments!\n");
-	else if (DBUS_TYPE_INT32 != dbus_message_iter_get_arg_type(&args))
-		fprintf(stderr, "Argument is not integer!\n");
+	if (!dbus_message_iter_init(msg, &args)) {
+		panic("%s", "Message has no arguments!\n");
+	} else if (DBUS_TYPE_INT32 != dbus_message_iter_get_arg_type(&args))
+		panic("%s", "Argument is not integer!\n");
 	dbus_message_iter_get_basic(&args, &qscale);
-	printf("qscale rate set:  %d\n", qscale);
+	debug(1, "%d", qscale);
 	send_ok(msg, conn);
 
 }
@@ -298,15 +287,13 @@ void qscale_get(DBusMessage * msg, DBusConnection * conn)
 
 	// add the arguments to the reply
 	dbus_message_iter_init_append(reply, &args);
-	if (!dbus_message_iter_append_basic(&args, DBUS_TYPE_INT32, &qscale)) 
-	{
-		fprintf(stderr, "Out Of Memory!\n");
+	if (!dbus_message_iter_append_basic(&args, DBUS_TYPE_INT32, &qscale)) {
+		panic("%s", "Out Of Memory!\n");
 		exit(1);
 	}
 	// send the reply && flush the connection
-	if (!dbus_connection_send(conn, reply, &serial)) 
-	{
-		fprintf(stderr, "Out Of Memory!\n");
+	if (!dbus_connection_send(conn, reply, &serial)) {
+		panic("%s", "Out Of Memory!\n");
 		exit(1);
 	}
 	dbus_connection_flush(conn);
@@ -321,20 +308,18 @@ void client_ip_set(DBusMessage * msg, DBusConnection * conn)
 	DBusMessageIter args;
 
 	// read the arguments
-	if (!dbus_message_iter_init(msg, &args))
-		fprintf(stderr, "Message has no arguments!\n");
-	else if (DBUS_TYPE_STRING != dbus_message_iter_get_arg_type(&args))
-		fprintf(stderr, "Argument is not string!\n");
+	if (!dbus_message_iter_init(msg, &args)) {
+		panic("%s", "Message has no arguments!\n");
+	} else if (DBUS_TYPE_STRING != dbus_message_iter_get_arg_type(&args))
+		panic("%s", "Argument is not string!\n");
 	char *text;
 	dbus_message_iter_get_basic(&args, &text);
-	if (clip) 
-	{
+	if (clip) {
 		free(clip);
 	}
 	clip = strdup(text);
 
-	printf(clip);
-	printf("\n");
+	debug(1, "%s", clip);
 	send_ok(msg, conn);
 }
 
@@ -349,15 +334,13 @@ void client_ip_get(DBusMessage * msg, DBusConnection * conn)
 
 	// add the arguments to the reply
 	dbus_message_iter_init_append(reply, &args);
-	if (!dbus_message_iter_append_basic(&args, DBUS_TYPE_STRING, &clip)) 
-	{
-		fprintf(stderr, "Out Of Memory!\n");
+	if (!dbus_message_iter_append_basic(&args, DBUS_TYPE_STRING, &clip)) {
+		panic("%s", "Out Of Memory!\n");
 		exit(1);
 	}
 	// send the reply && flush the connection
-	if (!dbus_connection_send(conn, reply, &serial)) 
-	{
-		fprintf(stderr, "Out Of Memory!\n");
+	if (!dbus_connection_send(conn, reply, &serial)) {
+		panic("%s", "Out Of Memory!\n");
 		exit(1);
 	}
 	dbus_connection_flush(conn);
@@ -372,9 +355,8 @@ void do_introspect(DBusMessage * msg, DBusConnection * conn)
 	DBusMessage *reply;
 	dbus_uint32_t serial = 0;
 	if (dbus_message_is_method_call
-	    (msg, "org.freedesktop.DBus.Introspectable", "Introspect")) 
-	{
-		printf("must do introspection \n");
+	    (msg, "org.freedesktop.DBus.Introspectable", "Introspect")) {
+		debug(2, "%s", "must do introspection");
 		static const char *introspect_xml =
 		    "<!DOCTYPE node PUBLIC \"-//freedesktop//DTD D-BUS Object Introspection 1.0//EN\"\n"
 		    "                      \"http://www.freedesktop.org/standards/dbus/1.0/introspect.dtd\">\n"
@@ -425,15 +407,13 @@ void do_introspect(DBusMessage * msg, DBusConnection * conn)
 		reply = dbus_message_new_method_return(msg);
 		if (!dbus_message_append_args(reply,
 					      DBUS_TYPE_STRING, &introspect_xml,
-					      DBUS_TYPE_INVALID)) 
-		{
-			fprintf(stderr, "Out Of Memory!\n");
+					      DBUS_TYPE_INVALID)) {
+			panic("%s", "Out Of Memory!\n");
 			exit(1);
 		}
 
-		if (!dbus_connection_send(conn, reply, &serial)) 
-		{
-			fprintf(stderr, "Out Of Memory!\n");
+		if (!dbus_connection_send(conn, reply, &serial)) {
+			panic("%s", "Out Of Memory!\n");
 			exit(1);
 		}
 		dbus_connection_flush(conn);
@@ -445,67 +425,54 @@ void do_introspect(DBusMessage * msg, DBusConnection * conn)
 void test_message(DBusMessage * msg, DBusConnection * conn)
 {
 	if (dbus_message_is_method_call
-	    (msg, "ch.cett.misse.ffmpeg", "image_rate_get")) 
-	{
+	    (msg, "ch.cett.misse.ffmpeg", "image_rate_get")) {
 		image_rate_get(msg, conn);
 	}
-	if (dbus_message_is_method_call(msg, "ch.cett.misse.ffmpeg", "start")) 
-	{
+	if (dbus_message_is_method_call(msg, "ch.cett.misse.ffmpeg", "start")) {
 		start(msg, conn);
 	}
-	if (dbus_message_is_method_call(msg, "ch.cett.misse.ffmpeg", "stop")) 
-	{
+	if (dbus_message_is_method_call(msg, "ch.cett.misse.ffmpeg", "stop")) {
 		stop(msg, conn);
 	}
 	if (dbus_message_is_method_call
-	    (msg, "ch.cett.misse.ffmpeg", "image_size_set")) 
-	{
+	    (msg, "ch.cett.misse.ffmpeg", "image_size_set")) {
 		image_size_set(msg, conn);
 	}
 	if (dbus_message_is_method_call
-	    (msg, "ch.cett.misse.ffmpeg", "image_size_get")) 
-	{
+	    (msg, "ch.cett.misse.ffmpeg", "image_size_get")) {
 		image_size_get(msg, conn);
 	}
 	if (dbus_message_is_method_call
-	    (msg, "ch.cett.misse.ffmpeg", "image_rate_set")) 
-	{
+	    (msg, "ch.cett.misse.ffmpeg", "image_rate_set")) {
 		image_rate_set(msg, conn);
 	}
 	if (dbus_message_is_method_call
-	    (msg, "ch.cett.misse.ffmpeg", "bit_rate_set")) 
-	{
+	    (msg, "ch.cett.misse.ffmpeg", "bit_rate_set")) {
 		bit_rate_set(msg, conn);
 	}
 	if (dbus_message_is_method_call
-	    (msg, "ch.cett.misse.ffmpeg", "bit_rate_get")) 
-	{
+	    (msg, "ch.cett.misse.ffmpeg", "bit_rate_get")) {
 		bit_rate_get(msg, conn);
 	}
 	if (dbus_message_is_method_call
-	    (msg, "ch.cett.misse.ffmpeg", "qscale_set")) 
-	{
+	    (msg, "ch.cett.misse.ffmpeg", "qscale_set")) {
 		qscale_set(msg, conn);
 	}
 	if (dbus_message_is_method_call
-	    (msg, "ch.cett.misse.ffmpeg", "qscale_get")) 
-	{
+	    (msg, "ch.cett.misse.ffmpeg", "qscale_get")) {
 		qscale_get(msg, conn);
 	}
 	if (dbus_message_is_method_call
-	    (msg, "ch.cett.misse.ffmpeg", "client_ip_set")) 
-	{
+	    (msg, "ch.cett.misse.ffmpeg", "client_ip_set")) {
 		client_ip_set(msg, conn);
 	}
 	if (dbus_message_is_method_call
-	    (msg, "ch.cett.misse.ffmpeg", "client_ip_get")) 
-	{
+	    (msg, "ch.cett.misse.ffmpeg", "client_ip_get")) {
 		client_ip_get(msg, conn);
 	}
 
 	return;
 }
-
 
 void listen()
 {
@@ -513,73 +480,56 @@ void listen()
 	DBusConnection *conn;
 	DBusError err;
 	int ret;
-	printf("Listening for method calls\n");
+	debug(1, "%s", "Listening for method calls\n");
 
 	// initialise the error
 	dbus_error_init(&err);
 
 	// connect to the bus and check for errors
 	conn = dbus_bus_get(DBUS_BUS_SESSION, &err);
-	if (dbus_error_is_set(&err)) 
-	{
-		fprintf(stderr, "Connection Error (%s)\n", err.message);
+	if (dbus_error_is_set(&err)) {
+		error("%s", "Connection Error");
+		error("%s", err.message);
 		dbus_error_free(&err);
 	}
-	if (NULL == conn) 
-	{
-		fprintf(stderr, "Connection Null\n");
+	if (NULL == conn) {
+		panic("%s", "Connection Null\n");
 		exit(1);
 	}
 	// request our name on the bus and check for errors
 	ret = dbus_bus_request_name(conn, "ch.cett.misse.ffmpeg",
 				    DBUS_NAME_FLAG_REPLACE_EXISTING, &err);
-	if (dbus_error_is_set(&err)) 
-	{
-		fprintf(stderr, "Name Error (%s)\n", err.message);
+	if (dbus_error_is_set(&err)) {
+		error("%s", err.message);
 		dbus_error_free(&err);
 	}
-	if (DBUS_REQUEST_NAME_REPLY_PRIMARY_OWNER != ret) 
-	{
-		fprintf(stderr, "Not Primary Owner (%d)\n", ret);
+	if (DBUS_REQUEST_NAME_REPLY_PRIMARY_OWNER != ret) {
+		panic("%s", "Not Primary Owner");
+		panic("%d", ret);
 		exit(1);
 	}
 	// loop, testing for new messages
-	while (true) 
-	{
+	while (true) {
+		if (status == 0) {
+			close(pfderr[1]);
+			close(pfdout[1]);
+			if (((read(pfderr[0], bufferr, BUFSIZ) != 0)
+			     || (read(pfdout[0], buffout, BUFSIZ) != 0))) {
+
+				if (bufferr != NULL) {
+					debug(1, "%s", bufferr);
+				} else {
+					debug(1, "%s", buffout);
+				}
+
+			}
+		}
 		// non blocking read of the next available message
 		waitpid(pid, &status, 1);
 		dbus_connection_read_write(conn, 0);
 		msg = dbus_connection_pop_message(conn);
-
-		if (status == 0) 
-		{
-			close(pfderr[1]);
-			close(pfdout[1]);
-			while (((read(pfderr[0], bufferr, BUFSIZ) != 0)
-				|| (read(pfdout[0], buffout, BUFSIZ) != 0))
-			       && (status == 0)) 
-			{
-				//printf("status est %d \n", status);
-				dbus_connection_read_write(conn, 0);
-				msg = dbus_connection_pop_message(conn);
-
-				// loop again if we haven't got a message
-				if (NULL != msg) 
-				{
-					do_introspect(msg, conn);
-					test_message(msg, conn);
-				}
-
-				if (bufferr != NULL)
-					printf("%s \n", bufferr);
-				else
-					printf("%s \n", buffout);
-
-			}
-		}
 		// loop again if we haven't got a message
-		if (NULL == msg) 
-		{
+		if (NULL == msg) {
 			sleep(1);
 			continue;
 		}
